Reorganizing
Put each class into their own script and fixed some dependencies
