+++ Assets/Scripts/ClassStructures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb213a8414431f042aad03719555585c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
