+++ Assets/Scripts/ClassStructures/Class.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86bfc1720836fce478ad9c4be03a5476
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
