+++ Assets/Scripts/ClassStructures/CreatedChar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 396d83e5eb487ed4fb9e11ace82b9651
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
