+++ Assets/Scripts/ClassStructures/Race.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc2246ab2d0d7434987c712858bfe2cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
