+++ Assets/Scripts/ClassStructures/Stats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: caa6fdd8465c3dd44b475b1fd2ede99e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
