+++ Assets/Scripts/ClassStructures/Trait.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using AdventuresOfOldMultiplayer;
+using Unity.Netcode;
+using UnityEngine.UI;
+using TMPro;
+
+public class Trait {
+    private Stats stats;
+    private string name;
+}
