+++ Assets/Scripts/ClassStructures/Trait.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1dd49d4140e74304e84c96012863a957
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
