+++ Assets/Scripts/Manager Scripts/CharManUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d50ac090656951f4b9e065291ac5ea8c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
